Added foreground subscriptions worker

+++ app/src/main/java/me/kofesst/android/moneyapp/model/AssetEntity.kt
@@ -9,7 +9,7 @@ import java.io.Serializable
 
 @Entity(tableName = "assets")
 data class AssetEntity(
-    @PrimaryKey(autoGenerate = true) val assetId: Int = 0,
+    @PrimaryKey(autoGenerate = true) val assetId: Long = 0,
     var name: String = "",
     var balance: Double = 0.0,
     var type: Int = AssetTypes.CARD.ordinal
